Add Adress Regex to characteristics
I believe, that it would be useful for developers to have the Regular Expressions under the characteristics category to check whether or not the Address a user added is actually valid. 
If e.g. a Blockexplorer/Developer/Wallet uses an API call/tries to create one, checking for a valid address beforehand would be nice to have.

I would propose doing the same for hashes.